rebase and move id_generator
