+++ tests/sentry/sentry_metrics/test_cloudspanner.py
@@ -1,7 +1,7 @@
 import pytest
 
 from sentry.sentry_metrics.indexer.cloudspanner.cloudspanner import CloudSpannerIndexer, IdCodec
-from sentry.sentry_metrics.indexer.cloudspanner.id_generator import get_id
+from sentry.sentry_metrics.indexer.id_generator import get_id
 
 
 @pytest.mark.parametrize(
